refactor: dirs and files
