+++ src/renderer/modules/slides/bg-cover.css
