+++ src/renderer/modules/slides/bg-cover.js
