+++ src/renderer/modules/slides/get-slides.js
