+++ src/renderer/modules/slides/main-frame.js
