+++ src/renderer/modules/slides/render.js
