+++ src/renderer/modules/slides/slide-element.js
