+++ src/renderer/modules/slides/slides.css
